requirements.txt
streamlit>=1.30.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
